FContext: Added action "relabel" which contains restorecon
Action is triggered by actions that used to contain restorecon
Also moved into into the provider
